dilermano_agent_crewai/pyproject.toml
[project]
name = "dilermano-scheduling-agent"
version = "0.1.0"
description = "Dilermano's scheduling agent for the A2A project."
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.5",
    "crewai[tools]>=0.126.0",
    "pydantic",
    "python-dotenv",
    "uvicorn",
    "google-generativeai",
] 
